Basic HAL module
