+++ cmake/OpenCVModule.cmake
@@ -176,6 +176,11 @@ macro(ocv_add_module _name)
       endif()
     endif()
 
+    # add HAL as dependency
+    if(NOT "${the_module}" STREQUAL "opencv_hal")
+      ocv_add_dependencies(${the_module} opencv_hal)
+    endif()
+
     # add self to the world dependencies
     if((NOT DEFINED OPENCV_MODULE_IS_PART_OF_WORLD
         AND NOT OPENCV_MODULE_${the_module}_CLASS STREQUAL "BINDINGS"
@@ -513,6 +518,18 @@ macro(ocv_include_modules)
       endif()
     elseif(EXISTS "${d}")
       ocv_include_directories("${d}")
+    endif()
+  endforeach()
+endmacro()
+
+# same as previous but with dependencies
+macro(ocv_include_modules_recurse)
+  ocv_include_modules(${ARGN})
+  foreach(d ${ARGN})
+    if(d MATCHES "^opencv_" AND HAVE_${d} AND DEFINED OPENCV_MODULE_${d}_DEPS)
+      foreach (sub ${OPENCV_MODULE_${d}_DEPS})
+        ocv_include_modules(${sub})
+      endforeach()
     endif()
   endforeach()
 endmacro()
