+++ modules/hal/src/precomp.hpp
@@ -1,0 +1,2 @@
+
+#include "opencv2/hal.hpp"
